third_party/WebKit/LayoutTests/fast/events/touch/gesture/pad-gesture-fling-expected.txt

Tests basic use of GestureFlingStart

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS actualWheelEventsOccurred is >= 2
PASS cumulativeScrollX is >= 300
PASS cumulativeScrollY is >= 300
PASS successfullyParsed is true

TEST COMPLETE
